Rename and relocate Linuz GTK project
