+++ PlatformSpecific/CodeHubX.Linux/CodeHubX.Linux.GTK/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("CodeHubX.GTK")]
+[assembly: AssemblyTitle("CodeHubX for Linux (GTK)")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("CodeHubX.GTK")]
-[assembly: AssemblyCopyright("Copyright ©  2018")]
+[assembly: AssemblyProduct("CodeHubX")]
+[assembly: AssemblyCopyright("Copyright © SICS, 2018")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
